Pipelined opengl
We are now working with a pipeline
That is each stage renders to a framebuffer (or gives some other output)
This framebuffer (or other output) is then inputted into the next stage
that then again renders to a framebuffer or other output.

We have a special stage the `DrawFramebufferStage`.
It takes a framebuffer and outputs it on the screen.
Thus it has no output.